added to SOSIX

+++ libs/sosix/stdlib/memory/free.c
@@ -9,7 +9,7 @@ void free(void *ptr) {
 	proc_control_block current_task = task_queue[0];
 
 	//lets compare addresses 
-	for (int i = 0; i <= current_task.heap_allocations_used; i++) {
+	for (size_t i = 0; i <= current_task.heap_allocations_used; i++) {
 		if (current_task.heap_allocations[i] == ptr) {
 			//lets free it
 			current_task.heap_allocations[i] = NULL;
